lab-report-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
python-consul2==0.1.5
